pyproject.toml
[project]
name = "etl-velo"
version = "0.6.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.10.11",
    "dbt-duckdb>=1.9.6",
    "minio==7.2.20",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "requests>=2.32.5",
    "streamlit>=1.50",
]

[dependency-groups]
dev = [
    "ruff>=0.13.0",
]

[tool.ruff.lint]
select = ["E","F","I"]